refactor: memo 요약 및 역량 분석 기능 chat gpi api 연결
