+++ .gitignore
@@ -41,3 +41,4 @@ src/main/resources/application-secret.yml
 src/main/resources/chat-prompt.txt
 src/main/resources/chat-summary-prompt.txt
 src/main/resources/ability-analysis-prompt.txt
+src/main/resources/memo-summary-prompt.txt
